requirements.txt
# Testing
pytest==8.3.4
pytest-playwright==0.5.2
playwright==1.49.1

# Reporting
allure-pytest==2.13.5

# Configuration
pyyaml==6.0.2
pydantic==2.10.3
pydantic-settings==2.6.1